fixed: Cleaned up MACDll build/clean. fixed: Ported MACDll to x86_64.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport@14491 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/cores/paplayer/MACDll/Source/MACLib/Old/APEDecompressOld.cpp
@@ -250,7 +250,7 @@ int CAPEDecompressOld::GetInfo(APE_DECOMPRESS_FIELDS Field, int nParam1, int nPa
             }
             else
             {
-                WAVEFORMATEX wfeFormat; GetInfo(APE_INFO_WAVEFORMATEX, (int) &wfeFormat, 0);
+                WAVEFORMATEX wfeFormat; GetInfo(APE_INFO_WAVEFORMATEX, (long) &wfeFormat, 0);
                 WAVE_HEADER WAVHeader; FillWaveHeader(&WAVHeader, 
                     (m_nFinishBlock - m_nStartBlock) * GetInfo(APE_INFO_BLOCK_ALIGN), 
                     &wfeFormat,    0);
